% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\name{click}
\alias{click}
\title{click}
\usage{
click()
}
\description{
This function adds or subtracts one from the counter depending on the tally counter type used as defined with the
tally_counter function. It prints the current count to the console with a minimum padding of four characters. Once
counter has finished counter is removed from its environment
}
\examples{
data <- data.frame(x = 1:10)
tally_counter(data, type = "add")
click()
}
