% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\name{tally_counter}
\alias{tally_counter}
\title{tally_counter}
\usage{
tally_counter(data, ...)
}
\arguments{
\item{data}{a data frame to be used in subsequent iteration}

\item{type}{character string indicating type of counter to use, either adding or subtracting counts}
}
\value{
the data frame is returned so that the function is pipe friendly
}
\description{
This function initiates a new tally counter with the count limit set to the number of rows in the data frame
 that will be used to interate over. The counter can either increase from zero to the number of rows set or
 decrease from the number of rows to zero.
}
\examples{
data <- data.frame(x = 1:10)
tally_counter(data, type = "add")
}
