% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tallyr-package.R
\docType{package}
\name{tallyr-package}
\alias{tallyr}
\alias{tallyr-package}
\title{tallyr: Import Form Data from 'Tally'}
\description{
An R interface to the 'Tally' form builder API \url{https://developers.tally.so}. Authenticate with an API key, list your forms, and import form submissions as tidy tibbles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EllaKaye/tallyr}
  \item Report bugs at \url{https://github.com/EllaKaye/tallyr/issues}
}

}
\author{
\strong{Maintainer}: Ella Kaye \email{hello@ellakaye.co.uk}

Authors:
\itemize{
  \item Ella Kaye \email{hello@ellakaye.co.uk}
}

}
\keyword{internal}
